+++ frontend/app/(auth)/login/components/GoogleLogin/hooks/useGoogleLogin.ts
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import { useState } from "react";
 
-import { useSupabase } from "@/app/supabase-provider";
+import { useSupabase } from "@/lib/context/SupabaseProvider";
 import { useToast } from "@/lib/hooks/useToast";
 
 export const useGoogleLogin = () => {
