+++ frontend/app/(auth)/signup/page.tsx
@@ -3,11 +3,11 @@
 import Link from "next/link";
 import { useState } from "react";
 
-import { useSupabase } from "@/app/supabase-provider";
 import Button from "@/lib/components/ui/Button";
 import Card from "@/lib/components/ui/Card";
 import Field from "@/lib/components/ui/Field";
 import PageHeading from "@/lib/components/ui/PageHeading";
+import { useSupabase } from "@/lib/context/SupabaseProvider";
 import { useToast } from "@/lib/hooks/useToast";
 
 export default function SignUp() {
