+++ frontend/app/explore/DocumentItem/DocumentData.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 
 import { useAxios } from "@/lib/useAxios";
 
-import { useSupabase } from "../../supabase-provider";
+import { useSupabase } from "../../../lib/context/SupabaseProvider";
 
 interface DocumentDataProps {
   documentName: string;
