+++ frontend/app/explore/page.tsx
@@ -7,10 +7,10 @@ import { useEffect, useState } from "react";
 
 import Button from "@/lib/components/ui/Button";
 import Spinner from "@/lib/components/ui/Spinner";
+import { useSupabase } from "@/lib/context/SupabaseProvider";
 import { Document } from "@/lib/types/Document";
 import { useAxios } from "@/lib/useAxios";
 
-import { useSupabase } from "../supabase-provider";
 import DocumentItem from "./DocumentItem";
 
 const ExplorePage = (): JSX.Element => {
