+++ frontend/app/upload/components/Crawler/hooks/useCrawler.ts
@@ -2,7 +2,7 @@
 import { redirect } from "next/navigation";
 import { useCallback, useRef, useState } from "react";
 
-import { useSupabase } from "@/app/supabase-provider";
+import { useSupabase } from "@/lib/context/SupabaseProvider";
 import { useToast } from "@/lib/hooks/useToast";
 import { useAxios } from "@/lib/useAxios";
