+++ frontend/app/user/page.tsx
@@ -7,7 +7,7 @@ import Spinner from "@/lib/components/ui/Spinner";
 import { UserStats } from "@/lib/types/User";
 import { useAxios } from "@/lib/useAxios";
 
-import { useSupabase } from "../supabase-provider";
+import { useSupabase } from "@/lib/context/SupabaseProvider";
 import { UserStatistics } from "./components/UserStatistics";
 
 const UserPage = (): JSX.Element => {
