+++ frontend/lib/context/SupabaseProvider/index.ts
@@ -1,0 +1,1 @@
+export * from "./supabase-provider";
